+++ app/src/main/java/nl/otters/elbho/adapters/ListAdapter.kt
@@ -11,8 +11,7 @@ import androidx.recyclerview.widget.RecyclerView
 import kotlinx.android.synthetic.main.fragment_listitem.view.*
 import nl.otters.elbho.R
 import nl.otters.elbho.models.Request
-import java.text.SimpleDateFormat
-import java.util.*
+import nl.otters.elbho.utils.DateParser
 import kotlin.collections.ArrayList
 
 
@@ -22,6 +21,7 @@ class ListAdapter(
     private val listener: OnClickItemListener
 //    private val bottomReachedListener: OnBottomReachedListener
 ) : RecyclerView.Adapter<ListAdapter.ViewHolder>(){
+    private val dateParser = DateParser()
 
     interface OnClickItemListener{
         fun onItemClick(position: Int, view: View)
@@ -50,53 +50,26 @@ class ListAdapter(
         val item = items[position]
 
         holder.titleView.text = item.cocName
-        holder.descriptionView.text = formatDescription(parseToTime(item.appointmentDatetime), item.address)
-        holder.dateView.text = parseToDate(item.appointmentDatetime)
-        holder.dayView.text = parseToDay(item.appointmentDatetime)
+        holder.descriptionView.text = formatDescription(item.appointmentDatetime, item.address)
+        holder.dateView.text = dateParser.toFormattedDate(item.appointmentDatetime)
+        holder.dayView.text = dateParser.toFormattedDay(item.appointmentDatetime)
         holder.icon.setImageResource(R.drawable.ic_chevron_right_24dp)
         holder.itemView.setOnClickListener{
             listener.onItemClick(holder.adapterPosition, it)
         }
     }
 
-    // TODO: put this in a util class
-    // TODO: do something with the Locale, like a factory or safe it in the sharedpref
-    private fun parseToDay(dateTime: String) : String{
-        val parser = SimpleDateFormat("yyyy-MM-dd'T'HH:mm:ss", Locale("nl"))
-        val formatter = SimpleDateFormat("EE",  Locale("nl"))
-        return formatter.format(parser.parse(dateTime)!!).toUpperCase()
-    }
-
-    private fun parseToDate(dateTime: String) : String{
-        val parser = SimpleDateFormat("yyyy-MM-dd'T'HH:mm:ss",  Locale("nl"))
-        val formatter = SimpleDateFormat("dd-MM",  Locale("nl"))
-        return formatter.format(parser.parse(dateTime)!!)
-    }
-
-    private fun parseToTime(dateTime: String) : String{
-        val parser = SimpleDateFormat("yyyy-MM-dd'T'HH:mm:ss",  Locale("nl"))
-        val formatter = SimpleDateFormat("HH:mm",  Locale("nl"))
-        return formatter.format(parser.parse(dateTime)!!)
-    }
-
     private fun formatDescription(startTime: String, address: String): String{
-        return startTime
+        val parsedStartTime: String = dateParser.toFormattedTime(startTime)
+        val endTime: String = dateParser.addHours(startTime, 1)
+        val parsedEndTime: String = dateParser.toFormattedTime(endTime)
+
+        return parsedStartTime
             .plus( " - ")
-            .plus(addHour(startTime, 1))
+            .plus(parsedEndTime)
             .plus( ", ")
             .plus(address.removeRange(address.indexOf(','), address.length)
         )
     }
-
-    private fun addHour(startTime: String, hoursToAdd: Int): String{
-        val parser = SimpleDateFormat("HH:mm",  Locale("nl"))
-        val date: Date = parser.parse(startTime)!!
-        val calendar: Calendar = Calendar.getInstance()
-        calendar.time = date
-        calendar.add(Calendar.HOUR, hoursToAdd)
-
-        return parser.format(calendar.time)
-    }
-
 }
